Fix type conversion compile error (issue #1397).
git-svn-id: https://chromiumembedded.googlecode.com/svn/branches/2171@1875 5089003a-bbd8-11dd-ad1f-f1f9622dbc98

+++ tests/unittests/frame_unittest.cc
@@ -1285,8 +1285,8 @@ int GetNavFromMultiNavURL(const std::string& url) {
 
 // Extract the origin from the URL.
 std::string GetOriginFromMultiNavURL(const std::string& url) {
-  const int pos = url.rfind("/");
-  EXPECT_TRUE(pos > 0);
+  const size_t pos = url.rfind("/");
+  EXPECT_TRUE(pos > 0U);
   return url.substr(0, pos + 1);
 }
